refactor: setup -> mpcsetup
