+++ backend/groth16/mpcsetup/marshal.go
@@ -1,4 +1,4 @@
-package setup
+package mpcsetup
 
 import (
 	"io"
