+++ backend/groth16/mpcsetup/setup.go
@@ -1,10 +1,10 @@
-package setup
+package mpcsetup
 
 import (
 	"github.com/consensys/gnark-crypto/ecc/bn254"
 	"github.com/consensys/gnark-crypto/ecc/bn254/fr/fft"
 	groth16 "github.com/consensys/gnark/backend/groth16/bn254"
-	"github.com/consensys/gnark/backend/groth16/setup/utils"
+	"github.com/consensys/gnark/backend/groth16/mpcsetup/utils"
 )
 
 func ExtractKeys(srs1 *Phase1, srs2 *Phase2, evals *Phase2Evaluations, nConstraints int) (pk groth16.ProvingKey, vk groth16.VerifyingKey) {
