+++ backend/groth16/mpcsetup/setup_test.go
@@ -1,4 +1,4 @@
-package setup
+package mpcsetup
 
 import (
 	"testing"
